mutil-app/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.5
gunicorn==20.1.0
requests==2.28.1
python-dotenv==0.21.0